[feature] budget

+++ app/src/main/java/com/example/fusmobilni/model/items/category/OfferingsCategory.java
@@ -1,17 +1,17 @@
 package com.example.fusmobilni.model.items.category;
 
 public class OfferingsCategory {
-    private final int id;
+    private final Long id;
     private String name;
     private String description;
 
-    public OfferingsCategory(int id, String name, String description) {
+    public OfferingsCategory(Long id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
